split validate into json validate and jwt validate
split validate into 2 sparate files as it was growing too big.
Now JWT processing is split from JSON validation
